requirements.txt
gunicorn
Flask
google-api-python-client
google-auth
google-auth-httplib2
google-auth-oauthlib
groq
pyTelegramBotAPI
python-dotenv
python-telegram
python-telegram-bot
Quart
requests
telegram
telegram_text